Ensure mem. order for output vol. matches input
This should fix a bug (see issue #2) where the output would appear
transposed if the input volume was stored in Fortran order.
If the volume and strel have different memory orders it is likely
that the output will still be unexpected. We could either reorder
the strel (potentially making a copy) or just raise a warning.